requirements (1).txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
reportlab==4.0.7
